wrote BaseModule class and feature modules are now subclasses of BaseModule
