+++ modular_bot/Module.py
@@ -1,0 +1,46 @@
+class BaseModule:
+    """
+    Base module class to be extended by feature modules.
+    """
+    command_char = ''
+    # To be updated in subclasses
+    module_name = ''
+    module_description = ''
+    commands = []
+
+    def __init__(self, user_cmd_char):
+        """
+        Sets command prefix while initializing.
+        :param user_cmd_char: command prefix.
+        """
+        self.command_char = user_cmd_char
+
+    def get_module_name(self):
+        """
+        Returns name of the module.
+        :return: name of the module in string.
+        """
+        return self.module_name
+
+    def get_module_description(self):
+        """
+        Returns description of the module.
+        :return: description of the module in string.
+        """
+        return self.module_description
+
+    def get_all_commands(self):
+        """
+        Returns all commands of the module.
+        :return: commands of the module in string list.
+        """
+        return self.commands
+
+    async def parse_command(self, bundle):
+        """
+        Decides which command should be executed and calls it.
+        To be implemented on subclasses.
+        :param bundle Dictionary passed in from caller.
+        :return: no return value.
+        """
+        pass
