+++ modular_bot/modules/lol_module.py
@@ -2,9 +2,10 @@ import logging
 import requests
 from bs4 import BeautifulSoup
 import discord
+from modular_bot.Module import BaseModule
 
 
-class LOLEsportsModule:
+class LOLEsportsModule(BaseModule):
     """
     Class for LOL e-sports Module. Gets player profile, top champions used, recent matches from best.gg.
     """
@@ -12,35 +13,6 @@ class LOLEsportsModule:
     module_description = "Fetches profile, top champions used or recent official matches of professional League of" \
                          " Legends player. All data referenced from best.gg."
     commands = ["lol_player", "lol_topchamps", "lol_recent"]
-    command_char = ''
-
-    def __init__(self, user_cmd_char):
-        """
-        Sets command prefix while initializing.
-        :param user_cmd_char: command prefix.
-        """
-        self.command_char = user_cmd_char
-
-    def get_module_name(self):
-        """
-        Returns name of the module.
-        :return: name of the module in string.
-        """
-        return self.module_name
-
-    def get_module_description(self):
-        """
-        Returns description of the module.
-        :return: description of the module in string.
-        """
-        return self.module_description
-
-    def get_all_commands(self):
-        """
-        Returns all commands of the module.
-        :return: commands of the module in string list.
-        """
-        return self.commands
 
     async def parse_command(self, bundle):
         """
